Enable SAL checking in core and bin for user mode builds
SAL in user mode is a bit more picky about tracing, along with its not handling a var int correctly

+++ src/core/crypto.c
@@ -497,7 +497,7 @@ QuicCryptoDiscardKeys(
         DiscardKeyType,
         Connection,
         "Discarding key type = %hhu",
-        KeyType);
+        (uint8_t)KeyType);
 
     QuicPacketKeyFree(Crypto->TlsState.WriteKeys[KeyType]);
     QuicPacketKeyFree(Crypto->TlsState.ReadKeys[KeyType]);
